fixed highlighted issues by reviewer
