+++ classroom.rb
@@ -1,6 +1,5 @@
 class Classroom
-  attr_accessor :label
-  attr_reader :students
+  attr_accessor :label, :students
 
   def initialize(label)
     @label = label
@@ -8,52 +7,7 @@ class Classroom
   end
 
   def add_student(student)
-    @students << student
     student.classroom = self
-  end
-end
-
-class Student
-  attr_accessor :name, :classroom
-
-  def initialize(name)
-    @name = name
-    @classroom = nil
-  end
-end
-
-class Book
-  attr_accessor :title, :author
-
-  def initialize(title, author)
-    @title = title
-    @author = author
-  end
-end
-
-class Rental
-  attr_accessor :date
-  attr_reader :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-    @book = book
-    @person = person
-    book.add_rental(self)
-    person.add_rental(self)
-  end
-end
-
-class Person
-  attr_accessor :name
-  attr_reader :rentals
-
-  def initialize(name)
-    @name = name
-    @rentals = []
-  end
-
-  def add_rental(rental)
-    @rentals << rental
+    @students << student
   end
 end
