+++ student.rb
@@ -1,13 +1,16 @@
 require_relative 'person'
+require_relative 'classroom'
 
-# Represents a student, inheriting from the Person class.
 class Student < Person
-  def initialize(id, age, classroom, parent_permission: true, name: 'Unknown')
-    super(id, age, parent_permission: parent_permission, name: name)
+  attr_accessor :classroom
+
+  def initialize(classroom, age, name = 'Unknown', parent_permission: true)
+    super(age, name, parent_permission: parent_permission)
     @classroom = classroom
+    classroom.students << self
   end
 
   def play_hooky
-    '¯\\(ツ)/¯'
+    '¯(ツ)/¯'
   end
 end
