+++ teacher.rb
@@ -1,9 +1,10 @@
 require_relative 'person'
 
-# Represents a teacher, inheriting from the Person class.
 class Teacher < Person
-  def initialize(id, age, specialization, parent_permission: true, name: 'Unknown')
-    super(id, age, parent_permission: parent_permission, name: name)
+  attr_reader :specialization
+
+  def initialize(specialization, id, age, name = 'Unknown', parent_permission: true)
+    super(id, age, name, parent_permission: parent_permission)
     @specialization = specialization
   end
