+++ trimmer_decorator.rb
